abstracting away implementation details of release notes serialization #54
