+++ src/main/groovy/org/mockito/release/internal/gradle/NotableReleaseNotesGeneratorTask.java
@@ -2,6 +2,7 @@ package org.mockito.release.internal.gradle;
 
 import org.gradle.api.DefaultTask;
 import org.gradle.api.tasks.TaskAction;
+import org.mockito.release.internal.gradle.util.DefaultReleaseNotesSerializer;
 import org.mockito.release.internal.gradle.util.ReleaseNotesSerializer;
 import org.mockito.release.notes.format.ReleaseNotesFormatters;
 import org.mockito.release.notes.model.ReleaseNotesData;
@@ -20,7 +21,7 @@ public class NotableReleaseNotesGeneratorTask extends DefaultTask {
     }
 
     @TaskAction public void generateReleaseNotes() {
-        ReleaseNotesSerializer releaseNotesSerializer = new ReleaseNotesSerializer(notesGeneration.getTemporarySerializedNotesFile());
+        ReleaseNotesSerializer releaseNotesSerializer = new DefaultReleaseNotesSerializer(notesGeneration.getTemporarySerializedNotesFile());
         Collection<ReleaseNotesData> releaseNotes = releaseNotesSerializer.deserialize();
         String notes = ReleaseNotesFormatters.notableFormatter(
                 notesGeneration.getIntroductionText(), notesGeneration.getDetailedReleaseNotesLink(), notesGeneration.getVcsCommitsLinkTemplate())
